Added doxygen config to the common lib.

+++ common/src/common/decoder.cpp
@@ -1,6 +1,12 @@
-//
-// Created by David Oberacker on 2019-07-31.
-//
+/**
+ * @file        decoder.cpp
+ * @author      David Oberacker (david.oberacker@gmail.com)
+ * @brief       Implementation of huffman decoder operations.
+ * @version     0.1
+ * @date        2019-08-26
+ * 
+ * @copyright   Copyright (c) 2019 David Oberacker
+ */
 
 #include <string>
 #include <map>
@@ -10,17 +16,41 @@
 
 #include <boost/dynamic_bitset.hpp>
 
-
 #include "common/common.hpp"
 
+/**
+ * @brief Struct describing a node in a huffman tree.
+ */
 struct Node
 {
+    /**
+     * @brief Symbol encoded in this node of the tree.
+     */
     uint8_t Symbol;
+
+    /**
+     * @brief Is this node a leaf?
+     */
     bool IsLeaf;
+
+    /**
+     * @brief Index of the left subtree. If node is a leaf, this is -1.
+     */
     int64_t Left;
+
+    /**
+     * @brief Index of the right subtree. If node is a leaf, this is -1.
+     */
     int64_t Right;
 };
 
+/**
+ * @brief Decodes a byte with a offset from a boost::dynamic_bitset.
+ * 
+ * @param data The bitset to use.
+ * @param offset The offset to use.
+ * @return uint8_t The byte that was read for the bitset.
+ */
 uint8_t decodeByte(boost::dynamic_bitset<> data, int64_t* offset) {
     uint8_t value = 0;
     for (int i = 7; i >= 0; --i)
@@ -31,7 +61,16 @@ uint8_t decodeByte(boost::dynamic_bitset<> data, int64_t* offset) {
     return value;
 };
 
-int decodeNode(const boost::dynamic_bitset<>& data, int64_t* offset, Node* tree, int64_t* index)
+/**
+ * @brief Function to decode a node in a encoded huffman tree.
+ * 
+ * @param[in] data Bitset containing the encoded data.
+ * @param[in] offset The offset from where to start decoding.
+ * @param[out] tree The generated tree, the decoded data should be added to.
+ * @param[out] index The index where to start adding the decoded nodes to.
+ * @return int64_t The decoded node index.
+ */
+int64_t decodeNode(const boost::dynamic_bitset<>& data, int64_t* offset, Node* tree, int64_t* index)
 {
     uint32_t current = (*index);
     (*index)++;
@@ -56,6 +95,13 @@ int decodeNode(const boost::dynamic_bitset<>& data, int64_t* offset, Node* tree,
     return current;
 }
 
+/**
+ * @brief Function to convert a byte sequence to a bitset.
+ * 
+ * @param[in] data The input byte sequence that should be converted.
+ * @param data_size The length of the input byte sequence. This should not be 0.
+ * @param[out] bit_map A pointer to the bitmap the results should be stored in.
+ */
  void convertBytesToBits(const char* data, uint32_t data_size, boost::dynamic_bitset<>* bit_map)
 {
     for (int i = 0, o = 0; i < data_size; i++)
@@ -67,8 +113,6 @@ int decodeNode(const boost::dynamic_bitset<>& data, int64_t* offset, Node* tree,
         }
     }
 }
-
-
 
 bool BORDERLANDS_COMMON_API
 D4v3::Borderlands::Common::Huffman::decode(const char *input_array, uint32_t input_size, char *output_array,
@@ -102,5 +146,4 @@ D4v3::Borderlands::Common::Huffman::decode(const char *input_array, uint32_t inp
 
     return true;
 }
-
